make meta collection generic and fix links
